+++ lib/occams-record/eager_loaders/has_one.rb
@@ -12,7 +12,13 @@ module OccamsRecord
       #
       def query(rows)
         return if rows.empty?
-        ids = rows.map { |r| r.send @ref.active_record_primary_key }.compact.uniq
+        ids = rows.map { |row|
+          begin
+            row.send @ref.active_record_primary_key
+          rescue NoMethodError => e
+            raise MissingColumnError.new(row, e.name)
+          end
+        }.compact.uniq
         q = base_scope.where(@ref.foreign_key => ids)
         q.where!(@ref.type => rows[0].class&.model_name) if @ref.options[:as]
         yield q if ids.any?